fix(home/emacs): stop corfu popupinfo opening frames for elisp
This might genuinely be the stupidest bug I have ever tracked down. Corfu uses
`elisp--company-doc-buffer` to get documentation, which was impossible to find,
because getting a backtrace was hell (I had to add a function to show raw
backtrace data to `trace-function-background` on `display-buffer` because I
could not find what was calling it). This function is even aware of its own
problems (comment in its source code), but its override of
`display-buffer-overriding-action` is inadequate, so now I have to override
`display-buffer-base-action` to get it to behave. Wow.