fix(home/emacs): use whitespace in wrap-prefix for custom indent
The previous version worked, but display properties would "leak" over the
indentation point; e.g. if a wrapped line began with a link, the link
styling (underline) would also show in the underlined region. This
implementation avoids that issue since the indented text now has the
`org-indent` face applied to it.