fix(home/emacs): use new meow motion setup
There's now a "passthrough" feature instead of the overriding setup from before.
Unfortunately, this means that if I want "u" to passthrough, I can no longer use
SPC u as a universal argument bind. So, instead, I'm using SPC l, which is the
next best thing ¯\_(ツ)_/¯ (it's where the QWERTY "u" key is).