feat(system/common/nix): remove Lix
Getting a build error on latest version, spent way too long trying to work
around it, went down the rabbit hole of older versions, and eventually gave up
on everything and decided it's not worth it. I don't even know what it's adding.
But perhaps I shall return in the future… but fornow, not worth it 😔. Perhaps
my days of bleeding-edge-everything are behind me now 🫡.